+++ libs/indibase/indiutility.cpp
@@ -1,0 +1,63 @@
+/*
+    Copyright (C) 2020 by Pawel Soja <kernel32.pl@gmail.com>
+    Copyright (C) 2015 by Jasem Mutlaq <mutlaqja@ikarustech.com>
+    Copyright (C) 2014 by geehalel <geehalel@gmail.com>
+
+    Stream Recorder
+
+    This library is free software; you can redistribute it and/or
+    modify it under the terms of the GNU Lesser General Public
+    License as published by the Free Software Foundation; either
+    version 2.1 of the License, or (at your option) any later version.
+
+    This library is distributed in the hope that it will be useful,
+    but WITHOUT ANY WARRANTY; without even the implied warranty of
+    MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.  See the GNU
+    Lesser General Public License for more details.
+
+    You should have received a copy of the GNU Lesser General Public
+    License along with this library; if not, write to the Free Software
+    Foundation, Inc., 51 Franklin Street, Fifth Floor, Boston, MA  02110-1301  USA
+
+*/
+#include "indiutility.h"
+
+namespace INDI
+{
+
+int mkpath(std::string s, mode_t mode)
+{
+    size_t pre = 0, pos;
+    std::string dir;
+    int mdret = 0;
+    struct stat st;
+
+    if (s[s.size() - 1] != '/')
+        s += '/';
+
+    while ((pos = s.find_first_of('/', pre)) != std::string::npos)
+    {
+        dir = s.substr(0, pos++);
+        pre = pos;
+        if (dir.size() == 0)
+            continue;
+
+        if (stat(dir.c_str(), &st))
+        {
+            if (errno != ENOENT || ((mdret = mkdir(dir.c_str(), mode)) && errno != EEXIST))
+            {
+                return mdret;
+            }
+        }
+        else
+        {
+            if (!S_ISDIR(st.st_mode))
+            {
+                return -1;
+            }
+        }
+    }
+    return mdret;
+}
+
+}
